Solve build issue on SolutionDir and mismatch between Release and Release_Win32
Add also support of french MSBuild error messages

+++ automated_build/msbuild_vsprojects.bat
@@ -33,20 +33,24 @@ if not exist %LOG_PATH% (
 	md %LOG_PATH%
 	echo log directory created
 ) else (
-	del /q %LOG_PATH%\*
-	echo old log files deleted
+	if not "%2"=="-no-delete" (
+		del /q %LOG_PATH%\*
+		echo old log files deleted
+	) else (
+		echo keeping old log files
+	)
 )
 
 if not exist %LOG_PASS% (
 	md %LOG_PASS%
 ) else (
-	del /q %LOG_PASS%\*
+	if not "%2"=="-no-delete" del /q %LOG_PASS%\*
 )
 
 if not exist %LOG_FAIL% (
 	md %LOG_FAIL%
 ) else (
-	del /q %LOG_FAIL%\*
+	if not "%2"=="-no-delete" del /q %LOG_FAIL%\*
 )
 
 if not exist %BIN_PATH% (
@@ -111,7 +115,7 @@ REM ----------------------------------------------------------------------------
 
 :LIB_X32
 	set BUILD=32
-	set FLAGS=/m /property:Configuration=Release
+	set FLAGS=/m /property:Configuration=Release /property:SolutionDir=%BATCH_PATH%..\
 	call :BUILD_PROJECT rx_common_lib, vcxproj
 	call :BUILD_PROJECT rx_common_lib_cs, csproj
 	call :BUILD_PROJECT rx_tif_lib, vcxproj
@@ -123,7 +127,7 @@ REM ----------------------------------------------------------------------------
 
 :BIN_X32
 	set BUILD=32
-	set FLAGS=/m /property:Configuration=Release
+	set FLAGS=/m /property:Configuration=Release /property:SolutionDir=%BATCH_PATH%..\
 	call :BUILD_PROJECT rx_main_ctrl, vcxproj
 	call :BUILD_PROJECT rx_spooler_ctrl, vcxproj
 	goto :EOF
@@ -131,7 +135,7 @@ REM ----------------------------------------------------------------------------
 
 :LIB_X64
 	set BUILD=64
-	set FLAGS=/m /property:Configuration=Release /property:Platform=x64
+	set FLAGS=/m /property:Configuration=Release /property:Platform=x64 /property:SolutionDir=%BATCH_PATH%..\
 	call :BUILD_PROJECT rx_common_lib, vcxproj
 	call :BUILD_PROJECT rx_common_lib_cs, csproj
 	call :BUILD_PROJECT TinyXML, vcxproj, Externals\
@@ -141,7 +145,7 @@ REM ----------------------------------------------------------------------------
 
 :BIN_X64
 	set BUILD=64
-	set FLAGS=/m /property:Configuration=Release /property:Platform=x64
+	set FLAGS=/m /property:Configuration=Release /property:Platform=x64 /property:SolutionDir=%BATCH_PATH%..\
 	call :BUILD_PROJECT rx_digiprint_gui, sln
 	call :BUILD_PROJECT Win10-Install, vcxproj, Win10\
 	call :BUILD_PROJECT Win10-Startup, vcxproj, Win10\
@@ -248,8 +252,8 @@ REM Parameter [3]: optional subpath of projectfile ending with '\'
 	REM get warnings and errors from logfile
 	set WARN_FILE=warn.tmp
 	set ERROR_FILE=error.tmp
-	powershell -Command "Select-String -Path %LOG_FILE% -Pattern '(\d*) (Warning\(s\)|Warnung\(en\))$' | %% { $_.Matches.groups[1] } | %% { $_.Value }" > %WARN_FILE%
-	powershell -Command "Select-String -Path %LOG_FILE% -Pattern '(\d*) (Error\(s\)|Fehler)$' | %% { $_.Matches.groups[1] } | %% { $_.Value }" > %ERROR_FILE%
+	powershell -Command "Select-String -Path %LOG_FILE% -Pattern '(\d*) (Warning\(s\)|Warnung\(en\)|Avertissement\(s\))$' | %% { $_.Matches.groups[1] } | %% { $_.Value }" > %WARN_FILE%
+	powershell -Command "Select-String -Path %LOG_FILE% -Pattern '(\d*) (Error\(s\)|Fehler|Erreur\(s\))$' | %% { $_.Matches.groups[1] } | %% { $_.Value }" > %ERROR_FILE%
 	set /p WARN= < %WARN_FILE%
 	set /p ERR= < %ERROR_FILE%
 	set WARN_STR=Warnings: %WARN%
